% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_online.R
\name{lda_online}
\alias{lda_online}
\title{Fit a Latent Dirichlet Allocation to text data}
\usage{
lda_online(dtm, passes = 1, batchsize = 256, K, eta = 1, alpha = 1,
  kappa = 0.7, tau_0 = 1024)
}
\arguments{
\item{dtm}{A data frame with three columns. The first column is a vector of document ids (a character vector), the second is a a vector of terms (a character vector), and the third is a vector of counts (an integer vector).}

\item{passes}{How many times we look at each document}

\item{batchsize}{The size of the minibatches}

\item{K}{The number of topics}

\item{eta}{hyperparameter}

\item{alpha}{hyperparameter}

\item{kappa}{learning rate parameter}

\item{tau_0}{learning rate parameter}
}
\description{
Fit a Latent Dirichlet Allocation to text data
}
