% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.svi-package.R
\docType{package}
\name{lda.svi-package}
\alias{lda.svi}
\alias{lda.svi-package}
\title{lda.svi: Fit Latent Dirichlet Allocation Models using Stochastic Variational Inference}
\description{
Fits Latent Dirichlet Allocation topic models to text data using the stochastic variational inference algorithm described in Hoffman et. al. (2013) <arXiv:1206.7051v3>. This method is more efficient than the original batch variational inference algorithm for LDA, and allows users to fit LDA models with more topics and to larger text corpora than would be feasible using that older method.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/nerskin/lda.svi/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Erskine \email{nicholas.erskine95@gmail.com}

}
\keyword{internal}
